Replace embedded script with pure CWL step
There are no commands to be executed here; the
step is purely a bookkeeping operation. To my
knowledge there is no strong advantage to using
a CommandLineTool or ExpressionTool; for both
no real operation needs to be executed that
cannot be done by pure CWL. POSIX has the colon
null operation, but this doesn't appear to be
supported by CWL.

I'm fairly sure that the current step is correct,
although the standard does not explicitly say that
the "writable: true" section should prevent any symlinking.