Allow prep_delay to execute TargetListToCoords.py
This commit removes some of the logic from prep_delay.cwl
and moves it to TargetListToCoords.py. The reasoning being
that the script can now operate independently of the workflow
without loss of function, which may help facilitate testing.

The CWL file still integrates the script into the larger workflow.