Add LINC as dependency to pyproject.toml
This will download the current LINC from its git repository,
and install in it as part of running `pip install .` from
 the VLBI root directory.

In other words, when installing the VLBI pipeline through pip,
it will automatically install LINC as a dependency. It will pull
in LINC's data files (skymodels, RFI strategies, etc), which
will be needed to run the delay calibration pipeline.

Ideally this is tied to some release of LINC, but the current
latest version (v5.1) has a version restriction of mocpy which
appears to prevent LINC installations through pip.