Use system installed LINC data files in setup workflow
I am assuming that the VLBI pipeline has been installed in
/usr/local/, such that LINC's scripts have been added to the
PATH and the data files have been placed in /usr/local/share/linc.

In this case, it is no longer needed to supply a local copy
of the LINC repository as a pipeline input. Instead, the
LINC scripts can be run directly on the command line, while
the data files can be accessed in from the path above.

I have implemented this for the setup workflow as a first step
in removing LINC as an input in favour of using it as a
pipeline dependency.