Add files via upload
Added the code quality_functions.py. Examples of usage are provided at the bottom of the code.

Given a 1" resolution image this code will provide 4 outputs which can be used to help ascertain image quality. Using the example at the bottom:

mode_fluxratio, std_fluxratio -- the mode and standard deivation fitted of the smearing corrected Total/Peak flux ratios. The mode_fluxratio would be 1.0 in the ideal case and increases as sources are more blurred.

fraction_matched, median_tot_fluxratio -- given an external catalogue and a noise mask (produced by the run_bdsf_fast routine) this sees how many sources may be detected in the 1" image. In the example dataset we find that when detecting sources in the high res map with a thresh_pix of 10 we recover 60% of the expected > 20sigma sources given the RMS map corrected for smearing. The median_tot_fluxratio is the flux density scale offset from LoTSS. Ideally this would be 1.0 but might be say 20-30% higher. The ratio is LoTSS/highres cat. 

If run with verbose lots of print statements and some plots are produced. If run without verbose then only the numbers are calculated.


Suggested tolerances for good/bad data.

mode_fluxratio < 1.5
median_tot_fluxratio < 1.5
fraction_matched > 0.5