Add differrent applycal opions for RM corrections
LINC has added support for `spinifex` in determining the rotation
measure. LINC target solsets will therefore no longer include an
RMextract table, but instead a spinifex one.

This is relevant for the A-team clipping workflow, where corrections
from this table are applied to the data in dp3_prep_target.cwl before
clipping. Here I allow for the possibility for the user to specify
which table must be used by adding an rm_corrections input to the
delay-calibration and setup workflows and the step dp3_make_parset.
This input specifies the table in LINC's solset that is to be used,
similar to the phase corrections in `phasesol`.

For now it is enforced that the input is either "RMextract" or
"spinifex".